+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ### Added
 
+- Command line arguments to disable individual steps
 - Functions and classes available at the top package level
 
 ### Fixed